Ful parallel numba support

+++ sleuth_sklearn/labeling.py
@@ -6,12 +6,12 @@ from numba import njit
 @njit
 def hoshen_kopelman(matrix):
     matrix = matrix.copy()
-    matrix = matrix.astype(np.int64)
+    matrix = matrix.astype(np.int32)
 
     m, n = matrix.shape
     max_labels = (m * n) // 2
 
-    labels = np.empty(max_labels, dtype=np.int64)
+    labels = np.empty(max_labels, dtype=np.int32)
     labels[0] = 0
 
     def find(x):
@@ -47,7 +47,7 @@ def hoshen_kopelman(matrix):
                 else:
                     matrix[i][j] = union(up, left)
 
-    new_labels = np.zeros(m * n // 2, dtype=np.int64)
+    new_labels = np.zeros(m * n // 2, dtype=np.int32)
     for i in range(m):
         for j in range(n):
             if matrix[i, j] > 0:
